bring back tests about LXCConfig
